requirements.txt
python-decouple==3.0
tweepy==3.5.0
